Change to root user (#11)
* Change to root user

* Add logging

* Add chown step to script

* Make dbt run as root

* Remove sudo (my bad) and set command for debug

* Add exit 1 for debug

* Restore dbt command

* Simplify inputs

The action can be simplified requiring the user to supply just one arguments.
The other inputs can be integrated as parameters of the argument itself, so
it is not required to explicitly pass them as separate inputs.

* Remove unnecessary logging

* Restore dbt-project input

the dbt project dir input has been restored. This will enable to cd
into the project directory and execute dbt commands that do not support
the --project-dir argument.

* Add git to docker image

dbt requires git to work properly.

* Remove git from image

Git has been installed in the docker image of pleo/dbt.